Fixed scalaStyle complaints: - Separated Services request messages from response ones in separated files   (scalaStyle was complaining of too much types declared in same file for RoomsCommon... was true...)
